==> templates/3_dataset.Rmd <==
```{r}
# Набор данных: Точность
df.acc <- read.csv("./data/accuracy_features.csv")

df.acc$acc_max[is.infinite(df.acc$acc_max)] <- NA
df.acc$acc_mean[is.infinite(df.acc$acc_mean)] <- NA
df.acc$acc_range[is.infinite(df.acc$acc_range)] <- NA

df.acc$dec_max[is.infinite(df.acc$dec_max)] <- NA
df.acc$dec_mean[is.infinite(df.acc$dec_mean)] <- NA
df.acc$dec_range[is.infinite(df.acc$dec_range)] <- NA

df.acc$X <- NULL
df.acc$stage <- "accuracy"

names(df.acc) <- tolower(names(df.acc))
```

```{r}
# Набор данных: Скорость
df.spd <- read.csv("./data/speed_features.csv")

df.spd$acc_max[is.infinite(df.spd$acc_max)] <- NA
df.spd$acc_mean[is.infinite(df.spd$acc_mean)] <- NA
df.spd$acc_range[is.infinite(df.spd$acc_range)] <- NA

df.spd$dec_max[is.infinite(df.spd$dec_max)] <- NA
df.spd$dec_mean[is.infinite(df.spd$dec_mean)] <- NA
df.spd$dec_range[is.infinite(df.spd$dec_range)] <- NA
df.spd$dec_min[is.infinite(df.spd$dec_min)] <- NA

df.spd$X <- NULL
df.spd$stage <- "speed"

names(df.spd) <- tolower(names(df.spd))
```

```{r}
## Данные тестов психодиагностики
sp_results <- read_xlsx("../../Все тесты/data/results/sp_results_impersonal.xlsx")

df.acc <- merge(df.acc, select(sp_results, userid, gender), by.x = "user_id", by.y = "userid", all.x = T)
df.acc <- select(df.acc, 1:4, gender, everything())
df.acc <- df.acc[!is.na(df.acc$gender),]
df.acc$gender <- factor(df.acc$gender, labels = c("F", "M"))

df.spd <- merge(df.spd, select(sp_results, userid, gender), by.x = "user_id", by.y = "userid", all.x = T)
df.spd <- select(df.spd, 1:4, gender, everything())
df.spd <- df.spd[!is.na(df.spd$gender),]
df.spd$gender <- factor(df.spd$gender, labels = c("F", "M"))
```

```{r}
## Кликеры
clickers.acc <- df.acc[df.acc$clicked_through_targets > 5 & df.acc$combo_clicked_through_targets > 5, ]
clickers.spd <- df.spd[df.spd$clicked_through_targets > 5 & df.spd$combo_clicked_through_targets > 5, ]
clickers <- unique(c(clickers.acc$user_id, clickers.spd$user_id))

df.acc <- df.acc[!df.acc$user_id %in% clickers, ]
df.spd <- df.spd[!df.spd$user_id %in% clickers, ]

save(clickers.acc, file = "./data/envs/1_desc_stat/clickers.acc.RData")
save(clickers.spd, file = "./data/envs/1_desc_stat/clickers.spd.RData")
```


```{r}
## outliers
## Выбросы найденные вручную с помощью кластеризации
outliers.acc <- df.acc[df.acc$user_id %in% c(1264, 2700, 1121, 1915, 2045, 2637, 1359, 1717, 1855, 2358, 2676, 1182, 1747, 1981, 2992, 2633), ]
outliers.spd <- df.spd[df.spd$user_id %in% c(1263, 1324, 1457, 2206), ]
outliers.diff <- df.acc[df.acc$user_id %in% c(1168, 1241, 1766, 2934), ] # Outliers найдены в diff
excluded_outliers <- unique(c(outliers.acc$user_id, outliers.spd$user_id, outliers.diff$user_id))

df.acc <- df.acc[!df.acc$user_id %in% excluded_outliers,]
df.spd <- df.spd[!df.spd$user_id %in% excluded_outliers,]

save(outliers.acc, file = "./data/envs/1_desc_stat/outliers.acc.RData")
save(outliers.spd, file = "./data/envs/1_desc_stat/outliers.spd.RData")

```


```{r}
# # Очень продолжительное выполнение теста
# longduration.acc <- df.acc[df.acc$time_total %in% boxplot.stats(df.acc$time_total, coef = 1)$out, ]
# longduration.spd <- df.spd[df.spd$time_total %in% boxplot.stats(df.spd$time_total, coef = 1)$out, ]
# longduration <- unique(c(longduration.acc$user_id, longduration.spd$user_id))
# 
# 
# df.acc <- df.acc[!df.acc$user_id %in% longduration, ]
# df.spd <- df.spd[!df.spd$user_id %in% longduration, ]

# save(longduration.acc, file = "./data/envs/1_desc_stat/longduration.acc.RData")
# save(longduration.spd, file = "./data/envs/1_desc_stat/longduration.spd.RData")
```


```{r}
# # Парадоксальные
# paradox.acc <- df.acc[df.acc$corr_sat < 0, ]
# paradox.spd <- df.spd[df.spd$corr_sat < 0, ]
# paradox <- unique(c(paradox.acc$user_id, paradox.spd$user_id))
# 
# 
# df.acc <- df.acc[!df.acc$user_id %in% paradox, ]
# df.spd <- df.spd[!df.spd$user_id %in% paradox, ]

# save(paradox.acc, file = "./data/envs/1_desc_stat/paradox.acc.RData")
# save(paradox.spd, file = "./data/envs/1_desc_stat/paradox.spd.RData")
```



```{r}
## Остановки во время выполнения теста (можно попробовать восстановить)
long_stops.acc <- df.acc[df.acc$time_max >= 10, ]
long_stops.spd <- df.spd[df.spd$time_max >= 10, ]
long_stops <- unique(c(long_stops.acc$user_id, long_stops.spd$user_id))

df.acc <- df.acc[!df.acc$user_id %in% long_stops, ]
df.spd <- df.spd[!df.spd$user_id %in% long_stops, ]
```

```{r}
## Длинный формат данных
df <- rbind(df.acc, df.spd)

df$gender <- factor(df$gender, labels = c("F", "M"))
df$stage <- factor(df$stage, levels = c("accuracy", "speed"))
df$user_id <- factor(df$user_id)
```


```{r}
save(df, file = "./data/envs/1_desc_stat/df.RData")
save(df.acc, file = "./data/envs/1_desc_stat/df.acc.RData")
save(df.spd, file = "./data/envs/1_desc_stat/df.spd.RData")
save(sp_results, file = "./data/envs/1_desc_stat/sp_results.RData")
```

```{r}
sign_differences <- vector()
outliers_list <- list()
```
