==> templates/3.1_ds_template.Rmd <==
---
title: "Описательная статистика"
author: "SVA"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    css: "../styles.css"
    code_folding: hide
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
    self_contained: no
    df_print: paged
  word_document:
    toc: yes
    toc_depth: '4'
editor_options:
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  include = TRUE,
  message = FALSE, 
  warning = FALSE
)

rm(list=ls())
source("functions.R")
library("ARTool")
```
<script>
function addClassToTitle(containerId, className) {
  // Находим контейнер по id
  var container = document.getElementById(containerId);
  if (!container) {
    console.warn('Контейнер с id "' + containerId + '" не найден.');
    return;
  }

  // Добавляем класс к найденному h3
  container.classList.add(className);
}
</script>


{{dataset}}


# Описательные статистики
{% for feature in data_dictionary %}
{% if feature.type == "numeric" %}
### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.name}}**

#### Descriptive stat

```{r}
data.frame(describe(df.{{stage}}${{feature.short_name}}, quant = c(0.25, 0.75)))
```


```{r fig.width=7, fig.height=4, results='asis'}
plot_title <- "{{feature.short_name}}"
plot_list <- draw_plots_continuous(df.{{stage}}, var = df.{{stage}}${{feature.short_name}}, title = plot_title, filename = NULL, binwidth = NULL, bins = 15, alpha = 0.5, angle = 0)

create_tabs(plot_list)
```



```{r fig.width=7, fig.height=4, results='asis'}
if (sum(df.{{stage}}${{feature.short_name}} <= 0, na.rm = T) > 0) {
  df.{{stage}}${{feature.short_name}}_log <- log1p(df.{{stage}}${{feature.short_name}})
} else {
  df.{{stage}}${{feature.short_name}}_log <- log(df.{{stage}}${{feature.short_name}})
}

plot_title <- "{{feature.short_name}}_log"
plot_list <- draw_plots_continuous(df.{{stage}}, var = df.{{stage}}${{feature.short_name}}_log, title = plot_title, filename = "log", binwidth = NULL, bins = 15, alpha = 0.5, angle = 0, tabname = "log_")

create_tabs(plot_list)
```


#### Normal test
```{r}
st <- shapiro.test(df.{{stage}}${{feature.short_name}});print(st)
check_normal_distribution(st)

ks <- ks.test(df.{{stage}}${{feature.short_name}}, "pnorm");print(ks)
check_normal_distribution(ks)
```

```{r results='asis'}
p1 <- ggplot(df.{{stage}}, aes(sample = {{feature.short_name}})) +  geom_qq(col=main_color) + geom_qq_line() + labs(title = "QQPlot - {{feature.short_name}}", tabname="QQPlot", filename=paste0(plot_title, "_QQplot"))

create_tabs(list(p1))
```


#### Outliers
```{r echo=FALSE}
outliers <- get_outliers(df.{{stage}}, df.{{stage}}${{feature.short_name}}, 1.5)
outliers <- select(outliers, 1:4, {{feature.short_name}}, everything())
outliers[order(outliers${{feature.short_name}}, decreasing = T),]
outliers[order(outliers${{feature.short_name}}),]

outliers_list${{feature.short_name}} <- outliers$user_id
```


{% elif feature.type == "character" %}
### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.short_name}} - {{feature.name}}**

```{r}
df.{{stage}}[, c("id", "{{feature.short_name}}")]
```

```{r}
table(df.{{stage}}${{feature.short_name}})
```

{% elif feature.type == "factor" %}

### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.name}}**

#### Tables

Распределение по переменной "{{feature.short_name}}"

```{r}
tab <- table(df.{{stage}}${{feature.short_name}})

data.frame(
  level = names(tab),
  frequency = as.vector(tab),
  percent = as.vector(round(prop.table(tab) * 100, 2))
)
```


{% else %}

{% endif %}

{% endfor %}

### Outliers

```{r}
save(outliers_list, file = "./data/envs/1_desc_stat/outliers.{{stage}}.RData")
```


### Корреляционный анализ {.tabset .tabset-fade .tabset-pills}

```{r}
corr_vars <- c("distance_trimmed", 
"dec_trimmed", 
"acc_q0.75", 
"pause_q0.25", 
"missings_range", 
"time_q0.75", 
"missings_se", 
"missings_max", 
"acc_q0.25", 
"dec_q0.25", 
"acc_per_sec_q0.25", 
"pause_se", 
"acc_per_sec_q0.75", 
"pause_kurtosis", 
"missings_q0.25", 
"pause_trimmed", 
"missings_q0.75", 
"acc_kurtosis", 
"missings_kurtosis", 
"dec_mad", 
"acc_mad", 
"pause_q0.75", 
"missings_median", 
"acc_per_sec_median", 
"time_median", 
"time_q0.25", 
"time_trend", 
"time_trimmed", 
"acc_per_sec_trimmed", 
"missings_trimmed", 
"time_mean", 
"distance_se", 
"velocity_se", 
"missings_mean", 
"distance_mean", 
"acc_range", 
"dec_range", 
"time_range", 
"dec_median", 
"time_max", 
"dec_kurtosis", 
"acc_trimmed", 
"pause_median", 
"velocity_q0.75", 
"distance_kurtosis", 
"distance_range", 
"pause_mad", 
"dec_q0.75", 
"velocity_q0.25", 
"acc_se", 
"acc_max", 
"acc_per_sec_mad", 
"time_mad", 
"dec_se", 
"time_kurtosis", 
"distance_max", 
"time_se", 
"missings_mad", 
"dec_min", 
"distance_median", 
"dec_sd", 
"distance_sd", 
"distance_q0.75", 
"acc_median", 
"velocity_mad", 
"velocity_range", 
"dec_mean", 
"time_sd", 
"pause_mean", 
"distance_q0.25", 
"acc_per_sec_se", 
"velocity_max", 
"missings_skew", 
"missings_min", 
"accuracy_rate", 
"time_min", 
"distance_min", 
"jerk_line", 
"velocity_mean", 
"missings_sd")
```

```{r}
df.{{stage}}.num <- df.{{stage}}[, sapply(df.{{stage}}, typeof) == "double"]

df.{{stage}}.num <- select(df.{{stage}}.num, -corr_vars, -ends_with("_log"))

cor_matrix <- cor(df.{{stage}}.num, method = "spearman", use = "pairwise.complete.obs")

cor_matrix[lower.tri(cor_matrix, diag = TRUE)] <- NA

cor_df <- as.data.frame(as.table(cor_matrix)) %>%
  filter(!is.na(Freq)) %>%
  arrange(desc(abs(Freq)))  # сортируем по абсолютному значению корреляции по убыванию
```

```{r fig.height=30, fig.width=30, results='asis'}
draw_corrplots(df.{{stage}}.num, method = "spearman", number.cex = 2, tl.cex = 3)
```
