==> templates/4_ds_by_group_template2.Rmd <==
---
title: "DS {{group}}"
author: "SVA"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    css: "../styles.css"
    code_folding: hide
    df_print: paged
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
    self_contained: no
  word_document:
    toc: yes
    toc_depth: '4'
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  include = TRUE,
  message = FALSE, 
  warning = FALSE
)

rm(list=ls())
source("functions.R")
library("ARTool")
```
<script>
function addClassToTitle(containerId, className) {
  // Находим контейнер по id
  var container = document.getElementById(containerId);
  if (!container) {
    console.warn('Контейнер с id "' + containerId + '" не найден.');
    return;
  }

  // Добавляем класс к найденному h3
  container.classList.add(className);
}
</script>

{{dataset}}


# Описательные статистики


# Описательные статистики

### Портрет {.tabset .tabset-fade .tabset-pills}
#### Основные показатели

```{r}
vars <- c('time_total', 'distance_total', 'pause_total', 'velocity_mean', 'missings_distance')
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```

#### Время
```{r}
vars <- grep("^time", names(df), value = T)
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```

#### Расстояние
```{r}
vars <- grep("^distance", names(df), value = T)
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```

#### Скорость
```{r}
vars <- grep("^velocity", names(df), value = T)
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```

#### Промахи
```{r}
vars <- grep("^missings|^accuracy", names(df), value = T)
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```

#### Паузы
```{r}
vars <- c("pause_total", "pause_number", "pause_mean", "pause_sd", "pause_median", "pause_trimmed", "pause_mad", "pause_min", "pause_max", "pause_range", "pause_skew", "pause_kurtosis", "pause_se", "pause_q0.25", "pause_q0.75", "pause_trend")
df.portrait <- select(df, vars, contains("clust"))

draw_cluster_portrait(df.portrait, vars, {{group}}, scale = "min_max")
```


### {{group}} {.tabset .tabset-fade .tabset-pills}

#### Table
```{r}
table(df${{group}})
```

#### Barplot
```{r}
ggplot(df, aes({{group}}, fill={{group}})) + geom_bar()
```


{% for feature in data_dictionary %}
{% if feature.type == "numeric" %}
### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.name}}**

#### Descriptive stat

```{r}
describeBy({{feature.short_name}} ~ {{group}}, data = df, quant = c(0.25, 0.75), mat = T)
```


```{r fig.width=7, fig.height=4, results='asis'}
plot_title <- "{{feature.short_name}}"
plot_list <- draw_plots_continuous_by_group(df, var = {{feature.short_name}}, df${{group}}, group_lab = "{{group}}", x_lab = "{{feature.name}}", title = plot_title)
create_tabs(plot_list)
```


#### Normal test
```{r}
st <- shapiro.test(df${{feature.short_name}});print(st)
check_normal_distribution(st)
ks <- ks.test(df${{feature.short_name}}, "pnorm");print(ks)
check_normal_distribution(ks)
```

```{r}
tryCatch(
    expr = {
      df %>%
        group_by({{group_name}}) %>%
        summarise(
          W = shapiro.test({{feature.short_name}})$statistic,
          p.value = shapiro.test({{feature.short_name}})$p.value,
          message = ifelse(p.value < 0.05, "Проверка на нормальность не пройдена", "Проверка на нормальность пройдена")
        )
    },
    error = function(e){
      print(e)
    }
)
```


```{r results='asis'}
p1 <- ggplot(df, aes(sample = {{feature.short_name}})) +  geom_qq(col=main_color) + geom_qq_line() + labs(title = "QQPlot - {{feature.short_name}}", tabname="QQPlot", filename=paste0(plot_title, "_QQplot")) + facet_wrap(". ~ {{group}}")

create_tabs(list(p1))
```

#### Сравнение групп

```{r}
users_with_na <- df$user_id[is.na(df${{feature.short_name}})]
df.test <- df[!df$user_id %in% users_with_na,]
df.test$user_id <- factor(df.test$user_id)

t_test_results <- pairwise.t.test(df.test${{feature.short_name}}, df.test${{group}}, p.adjust.method = "bonferroni");t_test_results
wilcox_test_results <- pairwise.wilcox.test(df.test${{feature.short_name}}, df.test${{group}}, p.adjust.method = "bonferroni");wilcox_test_results
```

```{r results='asis'}
if (sum(wilcox_test_results$p.value <= 0.05, na.rm = T) > 0) {
  sign_differences <- c(sign_differences, "{{feature.short_name}}")
  cat("<script>document.addEventListener('DOMContentLoaded', function() {addClassToTitle('{{feature.short_name}}', 'highlight');});</script>")
}
```

#### CI ~ {{group}}

```{r}
pd = position_dodge(0.1)
ggplot(df, aes(x={{group}}, y={{feature.short_name}}, color={{group}}, group={{group}})) +
  stat_summary(fun.data = mean_cl_boot, geom="errorbar", width=0.2, lwd=0.8, position=pd) +
  stat_summary(fun.data = mean_cl_boot, geom="line", size=1, position=pd) +
  stat_summary(fun.data = mean_cl_boot, geom="point", size=2.5, position = pd, pch = 15)

```


{% elif feature.type == "character" %}
### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.short_name}} - {{feature.name}}**

```{r}
df[, c("id", "{{feature.short_name}}")]
```

```{r}
table(df${{feature.short_name}})
```

{% elif feature.type == "factor" %}

### {{feature.short_name}} {.tabset .tabset-fade .tabset-pills}
**{{feature.name}}**

#### Tables

Распределение по переменной "{{feature.short_name}}"

```{r}
table(df${{feature.short_name}}, df${{group}})
```


{% else %}

{% endif %}

{% endfor %}


### Значимые различия
```{r results='asis'}
cat(paste0('<a href="#', sign_differences, '">', sign_differences, "</a><br>"))
```
