==> templates/5_dataset_sp_acc_female.Rmd <==

```{r}
load("./data/envs/4_cluster_analysis/df.acc.female.RData")
df$gender <- factor(df$gender, labels = "F")

# df.acc <- df.acc[df.acc$gender == "F", ]
```


```{r}
sp_results <- select(sp_results, userid, starts_with("BFST"), starts_with("MROT"), starts_with("KOT"), starts_with("VIT"), starts_with("LC"), starts_with("SOA"))
names(sp_results) <- tolower(names(sp_results))
```

```{r}
df <- merge(df, sp_results, by.x = "user_id", by.y = "userid")
```
